Add statistics for competition team
